Enemy ships now fire projectiles, player can die

+++ structs.h
@@ -26,8 +26,8 @@ typedef struct Entity
     int w;
     int h;
 
-    int dx;
-    int dy;
+    float dx;
+    float dy;
     int health;
     int reload;
     int side;
